refactor: Simplify loop condition in main.rs

+++ src/main.rs
@@ -31,7 +31,7 @@ fn main() {
         chunk_size = payload_size;
     }
 
-    while cursor < payload_size as u64 {
+    while cursor < payload_size {
         if cursor + chunk_size > payload_size {
             chunk_size -= cursor + chunk_size - payload_size;
         }
